requirements.txt
streamlit
nltk
sklearn
python-dotenv